IndividualAuth password resets: get user by email
Since the user is prompted for an email to reset their password, use
that to look up their user account, rather than the `UserName`
field. This makes a difference if you have decoupled those two fields,
allowing the user to specify their username rather than storing the
email to both.